🎨 restructure types
