+++ modules/types/src/app.ts
@@ -1,10 +1,117 @@
-import { BigNumber } from "ethers/utils";
-
-import { Address, CFCoreTypes, OutcomeType, SolidityValueType } from "./cf";
+import { Address, BigNumber, BigNumberish, SolidityValueType } from "./basic";
 import { CFCoreChannel } from "./channel";
+import {
+  MultiAssetMultiPartyCoinTransferInterpreterParams,
+  OutcomeType,
+  SingleAssetTwoPartyCoinTransferInterpreterParams,
+  SupportedNetwork,
+  TwoPartyFixedOutcomeInterpreterParams, 
+} from "./contracts";
 
 ////////////////////////////////////
-////// APP REGISTRY
+////// App Instances
+
+export type AppIdentity = {
+  channelNonce: BigNumberish;
+  participants: string[];
+  appDefinition: string;
+  defaultTimeout: number;
+};
+
+export type AppInterface = {
+  addr: string;
+  stateEncoding: string;
+  actionEncoding: string | undefined;
+};
+
+export type SignedStateHashUpdate = {
+  appStateHash: string;
+  versionNumber: number;
+  timeout: number;
+  signatures: string[];
+};
+
+export type AppABIEncodings = {
+  stateEncoding: string;
+  actionEncoding: string | undefined;
+};
+
+export type AppInstanceInfo = {
+  identityHash: string;
+  appDefinition: string;
+  abiEncodings: AppABIEncodings;
+  initiatorDeposit: BigNumber;
+  initiatorDepositTokenAddress: string;
+  responderDeposit: BigNumber;
+  responderDepositTokenAddress: string;
+  timeout: BigNumber;
+  proposedByIdentifier: string; // xpub
+  proposedToIdentifier: string; // xpub
+  intermediaryIdentifier?: string;
+  // Interpreter-related Fields:
+  twoPartyOutcomeInterpreterParams?: TwoPartyFixedOutcomeInterpreterParams;
+  multiAssetMultiPartyCoinTransferInterpreterParams?:
+    MultiAssetMultiPartyCoinTransferInterpreterParams;
+  singleAssetTwoPartyCoinTransferInterpreterParams?:
+    SingleAssetTwoPartyCoinTransferInterpreterParams;
+};
+
+export type AppInstanceJson = {
+  identityHash: string;
+  multisigAddress: string;
+  participants: string[];
+  defaultTimeout: number;
+  appInterface: AppInterface;
+  isVirtualApp: boolean;
+  appSeqNo: number;
+  latestState: SolidityValueType;
+  latestVersionNumber: number;
+  latestTimeout: number;
+  outcomeType: number;
+  // Derived from:
+  // contracts/funding/interpreters/TwoPartyFixedOutcomeInterpreter.sol#L10
+  twoPartyOutcomeInterpreterParams?: {
+    playerAddrs: [string, string];
+    amount: { _hex: string };
+    tokenAddress: string;
+  };
+  // Derived from:
+  // contracts/funding/interpreters/MultiAssetMultiPartyCoinTransferInterpreter.sol#L18
+  multiAssetMultiPartyCoinTransferInterpreterParams?: {
+    limit: { _hex: string }[];
+    tokenAddresses: string[];
+  };
+  singleAssetTwoPartyCoinTransferInterpreterParams?: {
+    limit: { _hex: string };
+    tokenAddress: string;
+  };
+};
+
+export type AppInstanceProposal = {
+  abiEncodings: AppABIEncodings;
+  appDefinition: string;
+  appSeqNo: number;
+  identityHash: string;
+  initialState: SolidityValueType;
+  initiatorDeposit: string;
+  initiatorDepositTokenAddress: string;
+  intermediaryIdentifier?: string;
+  outcomeType: OutcomeType;
+  proposedByIdentifier: string;
+  proposedToIdentifier: string;
+  responderDeposit: string;
+  responderDepositTokenAddress: string;
+  timeout: string;
+  // Interpreter-related Fields
+  twoPartyOutcomeInterpreterParams?: TwoPartyFixedOutcomeInterpreterParams;
+  multiAssetMultiPartyCoinTransferInterpreterParams?:
+    MultiAssetMultiPartyCoinTransferInterpreterParams;
+  singleAssetTwoPartyCoinTransferInterpreterParams?:
+    SingleAssetTwoPartyCoinTransferInterpreterParams;
+};
+
+////////////////////////////////////
+////// App Registry
 
 export const SupportedApplications = {
   CoinBalanceRefundApp: "CoinBalanceRefundApp",
@@ -13,16 +120,6 @@ export const SupportedApplications = {
   SimpleTwoPartySwapApp: "SimpleTwoPartySwapApp",
 };
 export type SupportedApplication = keyof typeof SupportedApplications;
-
-export const SupportedNetworks = {
-  ganache: "ganache",
-  goerli: "goerli",
-  homestead: "homestead",
-  kovan: "kovan",
-  rinkeby: "rinkeby",
-  ropsten: "ropsten",
-};
-export type SupportedNetwork = keyof typeof SupportedNetworks;
 
 export type DefaultApp = {
   actionEncoding?: string;
@@ -37,9 +134,8 @@ export type DefaultApp = {
 export type AppRegistry = DefaultApp[];
 
 ////////////////////////////////////
-////// APP TYPES
+// Generic Apps
 
-//////// General
 export type App<T = string> = {
   id: number;
   channel: CFCoreChannel;
@@ -81,19 +177,25 @@ export type AppStateBigNumber = AppState<BigNumber>;
 export type AppAction<T = string> = SimpleLinkedTransferAppAction | SolidityValueType;
 export type AppActionBigNumber = AppAction<BigNumber> | SolidityValueType;
 
-//////// Swap apps
+////////////////////////////////////
+// Swap Apps
+
 export type SimpleSwapAppState<T = string> = {
   coinTransfers: CoinTransfer<T>[][];
 };
 export type SimpleSwapAppStateBigNumber = SimpleSwapAppState<BigNumber>;
 
-//////// Simple transfer app
+////////////////////////////////////
+// Simple Transfer Apps
+
 export type SimpleTransferAppState<T = string> = {
   coinTransfers: CoinTransfer<T>[];
 };
 export type SimpleTransferAppStateBigNumber = SimpleTransferAppState<BigNumber>;
 
-//////// Simple linked transfer app
+////////////////////////////////////
+// Simple Linked Transfer Apps
+
 export type SimpleLinkedTransferAppState<T = string> = {
   coinTransfers: CoinTransfer<T>[];
   linkedHash: string;
@@ -107,7 +209,9 @@ export type SimpleLinkedTransferAppAction = {
   preImage: string;
 };
 
-////// Unidirectional transfer app
+////////////////////////////////////
+// Unidirectional Transfer Apps
+
 export type UnidirectionalTransferAppState<T = string> = {
   finalized: false;
   transfers: [CoinTransfer<T>, CoinTransfer<T>];
@@ -132,7 +236,9 @@ export enum UnidirectionalTransferAppStage {
   CHANNEL_CLOSED,
 }
 
-////// Unidirectional linked transfer app
+////////////////////////////////////
+// Unidirectional Linked Transfer Apps
+
 export type UnidirectionalLinkedTransferAppState<T = string> = {
   stage: UnidirectionalLinkedTransferAppStage;
   transfers: [CoinTransfer<T>, CoinTransfer<T>];
@@ -161,7 +267,9 @@ export enum UnidirectionalLinkedTransferAppStage {
   CHANNEL_CLOSED,
 }
 
-////// CoinBalanceRefund types
+////////////////////////////////////
+// CoinBalanceRefund
+
 export type CoinBalanceRefundAppState<T = string> = {
   multisig: string;
   recipient: string;
