+++ modules/types/src/client.ts
@@ -1,16 +1,12 @@
-import { Contract, providers } from "ethers";
-import { BigNumber, Network } from "ethers/utils";
-
+import { AppInstanceJson } from "./app";
 import {
   AppActionBigNumber,
   AppRegistry,
   AppState,
   DefaultApp,
   SupportedApplication,
-  SupportedNetwork,
 } from "./app";
-import { ConnextEvent } from "./basic";
-import { AppInstanceJson, CFCoreTypes } from "./cf";
+import { BigNumber, Contract, JsonRpcProvider, Network } from "./basic";
 import { CFCoreChannel, ChannelAppSequences, ChannelState, PaymentProfile } from "./channel";
 import {
   ChannelProviderConfig,
@@ -18,6 +14,8 @@ import {
   IChannelProvider,
   KeyGen,
 } from "./channelProvider";
+import { SupportedNetwork } from "./contracts";
+import { ConnextEvent } from "./events";
 import {
   CheckDepositRightsParameters,
   CheckDepositRightsResponse,
@@ -43,13 +41,14 @@ import {
   RequestCollateralResponse,
   Transfer,
 } from "./node";
+import { ProtocolTypes } from "./protocol";
 import { IAsyncStorage, Store } from "./store";
 
 export type InternalClientOptions = ClientOptions & {
   appRegistry: AppRegistry;
   channelProvider: IChannelProvider;
   config: GetConfigResponse;
-  ethProvider: providers.JsonRpcProvider;
+  ethProvider: JsonRpcProvider;
   messaging: IMessagingService;
   network: Network;
   node: INodeApiClient;
@@ -106,7 +105,7 @@ export interface IConnextClient {
   channelProviderConfig(): Promise<ChannelProviderConfig>;
   requestDepositRights(
     params: RequestDepositRightsParameters,
-  ): Promise<CFCoreTypes.RequestDepositRightsResult>;
+  ): Promise<ProtocolTypes.RequestDepositRightsResult>;
   rescindDepositRights(
     params: RescindDepositRightsParameters,
   ): Promise<RescindDepositRightsResponse>;
@@ -144,37 +143,37 @@ export interface IConnextClient {
 
   ///////////////////////////////////
   // CF MODULE EASY ACCESS METHODS
-  deployMultisig(): Promise<CFCoreTypes.DeployStateDepositHolderResult>;
-  getStateChannel(): Promise<CFCoreTypes.GetStateChannelResult>;
+  deployMultisig(): Promise<ProtocolTypes.DeployStateDepositHolderResult>;
+  getStateChannel(): Promise<ProtocolTypes.GetStateChannelResult>;
   providerDeposit(
     amount: BigNumber,
     assetId: string,
     notifyCounterparty: boolean,
-  ): Promise<CFCoreTypes.DepositResult>;
-  getFreeBalance(assetId?: string): Promise<CFCoreTypes.GetFreeBalanceStateResult>;
+  ): Promise<ProtocolTypes.DepositResult>;
+  getFreeBalance(assetId?: string): Promise<ProtocolTypes.GetFreeBalanceStateResult>;
   getAppInstances(multisigAddress?: string): Promise<AppInstanceJson[]>;
-  getAppInstanceDetails(appInstanceId: string): Promise<CFCoreTypes.GetAppInstanceDetailsResult>;
-  getAppState(appInstanceId: string): Promise<CFCoreTypes.GetStateResult>;
+  getAppInstanceDetails(appInstanceId: string): Promise<ProtocolTypes.GetAppInstanceDetailsResult>;
+  getAppState(appInstanceId: string): Promise<ProtocolTypes.GetStateResult>;
   getProposedAppInstances(
     multisigAddress?: string,
-  ): Promise<CFCoreTypes.GetProposedAppInstancesResult | undefined>;
+  ): Promise<ProtocolTypes.GetProposedAppInstancesResult | undefined>;
   getProposedAppInstance(
     appInstanceId: string,
-  ): Promise<CFCoreTypes.GetProposedAppInstanceResult | undefined>;
+  ): Promise<ProtocolTypes.GetProposedAppInstanceResult | undefined>;
   proposeInstallApp(
-    params: CFCoreTypes.ProposeInstallParams,
-  ): Promise<CFCoreTypes.ProposeInstallResult>;
-  installVirtualApp(appInstanceId: string): Promise<CFCoreTypes.InstallVirtualResult>;
-  installApp(appInstanceId: string): Promise<CFCoreTypes.InstallResult>;
-  rejectInstallApp(appInstanceId: string): Promise<CFCoreTypes.UninstallResult>;
+    params: ProtocolTypes.ProposeInstallParams,
+  ): Promise<ProtocolTypes.ProposeInstallResult>;
+  installVirtualApp(appInstanceId: string): Promise<ProtocolTypes.InstallVirtualResult>;
+  installApp(appInstanceId: string): Promise<ProtocolTypes.InstallResult>;
+  rejectInstallApp(appInstanceId: string): Promise<ProtocolTypes.UninstallResult>;
   takeAction(
     appInstanceId: string,
     action: AppActionBigNumber,
-  ): Promise<CFCoreTypes.TakeActionResult>;
+  ): Promise<ProtocolTypes.TakeActionResult>;
   updateState(
     appInstanceId: string,
     newState: AppState | any,
-  ): Promise<CFCoreTypes.UpdateStateResult>;
-  uninstallApp(appInstanceId: string): Promise<CFCoreTypes.UninstallResult>;
-  uninstallVirtualApp(appInstanceId: string): Promise<CFCoreTypes.UninstallVirtualResult>;
+  ): Promise<ProtocolTypes.UpdateStateResult>;
+  uninstallApp(appInstanceId: string): Promise<ProtocolTypes.UninstallResult>;
+  uninstallVirtualApp(appInstanceId: string): Promise<ProtocolTypes.UninstallVirtualResult>;
 }
