+++ modules/types/src/constants.ts
@@ -1,0 +1,2 @@
+// 25446 is 0x6366... or "cf" in ascii, for "Counterfactual".
+export const CF_PATH = "m/44'/60'/0'/25446";
