+++ modules/types/src/index.ts
@@ -1,13 +1,16 @@
 export * from "./app";
 export * from "./basic";
-export * from "./cf";
+export * from "./cfCore";
 export * from "./channel";
 export * from "./channelProvider";
 export * from "./client";
+export * from "./constants";
+export * from "./contracts";
 export * from "./convert";
+export * from "./events";
 export * from "./inputs";
 export * from "./messaging";
 export * from "./node";
-export * from "./pisaClient";
+export * from "./protocol";
 export * from "./state";
 export * from "./store";
