+++ modules/types/src/messaging.ts
@@ -1,4 +1,39 @@
-import { CFCoreTypes } from "./cf";
+import { EventName } from "./events";
+
+// The message type for Nodes to communicate with each other.
+export type NodeMessage = {
+  from: string;
+  type: EventName;
+};
+
+type JsonRpcProtocolV2 = {
+  jsonrpc: "2.0";
+};
+
+type RpcParameters =
+  | {
+      [key: string]: any;
+    }
+  | any[];
+
+export type JsonRpcNotification = JsonRpcProtocolV2 & {
+  result: any;
+};
+
+export type JsonRpcResponse = JsonRpcNotification & {
+  id: number;
+};
+
+export type Rpc = {
+  methodName: string;
+  parameters: RpcParameters;
+  id?: number;
+};
+
+export interface IRpcNodeProvider {
+  onMessage(callback: (message: JsonRpcResponse | JsonRpcNotification) => void): any;
+  sendMessage(message: Rpc): any;
+}
 
 export interface MessagingConfig {
   clusterId?: string;
@@ -7,11 +42,16 @@ export interface MessagingConfig {
   logLevel: number;
 }
 
-export interface IMessagingService extends CFCoreTypes.IMessagingService {
+export interface CFMessagingService {
+  send(to: string, msg: NodeMessage): Promise<void>;
+  onReceive(address: string, callback: (msg: NodeMessage) => void): any;
+}
+
+export interface IMessagingService extends CFMessagingService {
   connect(): Promise<void>;
   disconnect(): Promise<void>;
   flush(): Promise<void>;
-  onReceive(subject: string, callback: (msg: CFCoreTypes.NodeMessage) => void): Promise<void>;
+  onReceive(subject: string, callback: (msg: NodeMessage) => void): Promise<void>;
   publish(subject: string, data: any): Promise<void>;
   request(
     subject: string,
@@ -19,7 +59,7 @@ export interface IMessagingService extends CFCoreTypes.IMessagingService {
     data: object,
     callback?: (response: any) => any,
   ): Promise<any>;
-  send(to: string, msg: CFCoreTypes.NodeMessage): Promise<void>;
-  subscribe(subject: string, callback: (msg: CFCoreTypes.NodeMessage) => void): Promise<void>;
+  send(to: string, msg: NodeMessage): Promise<void>;
+  subscribe(subject: string, callback: (msg: NodeMessage) => void): Promise<void>;
   unsubscribe(subject: string): Promise<void>;
 }
