--- modules/types/src/pisaClient.ts
@@ -1,1 +1,0 @@
-export { PisaClient as IPisaClient } from "pisa-client";
